functions/process-image/requirements.txt
boto3
rasterio
rio-cogeo
pyarrow
geopandas
shapely
Pillow
